+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 
 ### Administrator
 
+- 🟢 [#2815](https://redmine.czechidm.com/issues/2815) - Profile image is validated now:
+  - default maximum image file size is ``512KB``. Change configuration property ``idm.sec.core.identity.profile.image.max-file-size`` if needed.
+_Example ``idm.sec.core.identity.profile.image.max-file-size=200KB``._
+
+### Developer
+
+- 🟢 [#2815](https://redmine.czechidm.com/issues/2815) - Profile image is validated now:
+  - ``image/*`` content types are supported only. Content type is required now, when images are uploaded by rest endpoint.
+
+## [11.1.0]
+
+### Administrator
+
 - 🟢 [#2815](https://redmine.czechidm.com/issues/2815) - Role catalogue item code is hidden in role catalogue tree by default (in role select and agendas). Set application property ``idm.pub.app.show.roleCatalogue.tree.code=true``, then code will be shown in tree again.
 
 ### Developer